test_statistics
